+++ CHANGELOG.md
@@ -4,14 +4,12 @@
 
 ### Added
 - Add `dx-data-grid-generic` storybook story
-- Add `iconUrl` property to the `dx-chip` component
+- Add `icon` property to the `dx-chip` component
 
 ### Fixed
 - Fix the `XS` size of the `dx-dialog` component
 
 ### Changed
-- Remove unused `error` property from the `dx-multiple-select-chip` component
-- Remove unused `color` property from the `dx-header` component
 - Check and rework all storybook stories
 - Adjust `dx-chip` stylings
 - Adjust `dx-datepicker` stylings
@@ -21,6 +19,9 @@
 - Adjust `dx-snackbar` stylings
 
 ### Breaking changes
+- Remove unused `error` property from the `dx-multiple-select-chip` component
+- Remove unused `color` property from the `dx-header` component
+- Change `dx-data-grid-generic` interface propertie types (isLoading, hasMiddlewareError, hasContentSourceAvailable, checkboxSelection)
 - Remove `dx-search-center-layout` component
 
 ## 1.4.0